hnjnmy/requirements.txt
beautifulsoup4
Django
lxml
mysqlclient
Pillow
PyMySQL
pytz
